requirements.txt
accelerate==1.11.0
datasets==4.2.0
einops==0.8.1
fire==0.7.1
hydra-core==1.3.2
ipdb
lm-eval==0.4.9
magic-wormhole
mteb==2.7.2
mosaicml==0.32.1
mosaicml-streaming==0.13.0
pre-commit
ruff
sentencepiece==0.2.1
# torch==2.5.1
# Upgrade torch due to:
#   ╰─▶ Because mosaicml==0.32.1 depends on torch>=2.6.0,<2.7.1 and you require mosaicml==0.32.1, we can conclude that you require torch>=2.6.0,<2.7.1.
#       And because you require torch==2.5.1, we can conclude that your requirements are unsatisfiable.
torch==2.7.0
transformers==4.52.4
wandb==0.22.2
